Bython Windows BAT
This BAT file makes running Bython even easier as you can set your Bython extension to point to this file for easy Compilation when double clicked. This file can also be ran alone to open a Bython terminal and added to your Environment Variables to access Bython anywhere with ease!